* Fixed output in C++ implementation.
Fixed the head writer class. Any double underscores are not allowed in header
guard definitions. This required an additional change in the header interface,
removing the ending extension added to normal namespaces that is not the global
namespace. This now makes the global namespace header guard unique without using
double underscores.

Updated the settings class. Removed the trailing underscore of the root define
parameter as it is no longer needed.