* Fixed output in C++ implementation.
Fixed the function block class. Fixed the key method; it would return a key with
leading semicolons if the function was part of the global namespace when it
should start with the function name in that case.

Fixed the head writer class. Fixed the header interface; it was not properly
updated for the new flat file output, leaving out the extension part of the
header guard definition if it was the root namespace.