docs: Updated contributing types.
docs(CONTRIBUTING): Removed the intern type. This was found to cause
more problems than solutions. It was meant to reduce the amount of
commit messages that would be scanned when making a change log for a new
version. It would end up censoring out too many commit messages that
would be useful to know the details of any new feature.

Github Issue #1
Copy Existing Code