test: Added C++ language test.
test(main): Refactored how argument handling is done to determine if
unit tests should be executed or the normal application, removing the
use of Qt command line parser. This was done because the command line
parser was conflicting with the Qt unit test system, both of which
process the command line arguments.

test(Test): Refactored the execute function. Only a single Qt test class
can now be executed for the life of the program because it was
discovered the Qt unit test system does not allow multiple Qt test
classes being executed during the life of a program. The first two
command line arguments are now expected to be the test option and its
value, the value being the test name to execute. If no value is set or
the value is not a valid test it prints a list of all available tests.

test(Test::Language::Cpp): Added C++ language test class. It unit tests
the C++ language class.

GitHub Issue #1
Copy Existing Code