test+: Fixed broken C++ unit tests.
test+(Test::Language::Cpp): Fixed test cases for create class and
namespace. Added test case for creating function blocks.

test(Test::Block::Cpp::Class): Updated test case for the display icon
property to account for the new dynamic changes that can occur as a
result of function block children.

fix(Block::Cpp::Class): Fixed issues regarding a class block not
realizing its display icon needs to be updated as a result of one of its
children function blocks being removed. This required the addition of an
ignore variable for the “is abstract/virtual” methods and the update
display icon push method because a class block needs to ignore a child
function block that is about to be removed.

fix(Block::Cpp::Function): Added missing logic in display icon getter
method to display static function icons when the function is a static
method.

perf(Block::Cpp::Function): Added internal getter method for the reverse
flag lookup hash, where it is built upon request for the first time and
then saved for any later request. This will reduce the chance for errors
in the future.

feature(data/cpp): Added missing invalid function icon image.

perf(
  Block::Cpp::Class
  Block::Cpp::Function
): Adapted how the display icon property is changed and gotten by saving
a pointer to the current display icon and constructing icons as they are
needed with static getter methods. As a result of this change a special
type of setter method was added for the property that takes a pointer to
a Qt icon.

refactor(Block::Cpp::Function): Renamed all “add ...” methods which add
to the display text and help construct a function’s scope name to
“append ...” so they are not confused with the add event interface.

fix(Block::Cpp::Function): The global class constant hash table “flag
strings” all had the same function flag for each line when they should
be every possible flag matching the string it is associated with. This
was a copy paste typo and it was fixed.

GitHub Issue #5
Add C++/Qt Language Blocks