refactor: Linked C++ property & class blocks.
refactor(Block::Cpp::Property): Moved all common methods and interface
implementations from the class C++ block to this block. This includes
testing if it is virtual or abstract, setting a dynamic display icon
with a pointer, and providing a method that forces the updating of its
display icon. The update display icon method is now virtual, allowing
the class block to override it with its own class icons. As a result of
moving this over to properties, properties now have icons for regular,
virtual, and abstract just like classes.

refactor(Block::Cpp::Class): Removed all deprecated interfaces and
methods which now reside in the property C++ block class. Changed
inheritance from the base C++ block class to the property C++ block
class.

refactor(Block::Cpp::Function): Updated all code that was interacting
with the class C++ block to the new common property C++ block which is
shared with the class C++ block. Added “is property method” method.

feature(data/cpp): Updated C++ property icon to be a little smaller.
Added C++ abstract and virtual property icons.

test(Test::Block::Cpp::Class): Updated display icon property test to
correctly register a display icon changed signal even though the
pointers are the same. This was incorrectly not updated before that
method was moved to the property C++ block class. Just because a pointer
to an object is the same does NOT mean it is the same object.

GitHub Issue #5
Add C++/Qt Language Blocks