feature: Added main window widget.
feature(Widget::Window::Main): Added new main window widget. This is the
main window for this application. It is a skeleton widget with very
little actually implemented because it mostly provides interfaces to the
other widgets that do not yet exist. In this first pass of coding the
menus and tool bars were implemented as much as possible. The file and
help menus were fully implemented. All actions currently do nothing with
“TO DO” tags.

feature(main): Integrated new main window, creating a new window and
showing it on normal application startup.

GitHub Issue #7
Add C++/Qt Language Widgets