feature: Added to project widget.
feature(Widget::Project): Added “add to global” menu, slot, and “update
add global actions” private method. These combined implement a secondary
add menu which adds the new block to the root, or global, block of a
project.

fix(Widget::Project): Removed ability of the regular add menu to add new
blocks to the root block. This ability was removed because it caused a
segmentation fault bug within Qt independent of this application’s
source code. The best guess of this commit author to the cause is that
the Qt tree view class does not like not having a valid Qt model index
and attempts to update it in the middle of calling the add lambda which
is being deleted beforehand.

feature(Block::Cpp::Class): Added list of a class block’s parents
property to its display text property.

GitHub Issue #7
Add C++/Qt Language Widgets