feature: Added to project widget.
feature(Widget::Project): Added block widget container widgets. This
includes a Qt scroll area within a Qt group box within a custom widget
layout with an apply button on the bottom. Updated the “set block
widget” method to correctly set a block widget and connect it with the
new apply button if it is an abstract block widget. Added code in the
“on current index changed” slot to update the block widget based off the
new index, creating a widget from the block in a project widget
instance’s project model if it is valid. This now results in the program
crashing since none of the blocks have implemented any of their
respective block widgets.

intern(Model::Project): Added “constant block” getter method which
returns a constant pointer to the block in a project with the given
index. This was required by the project widget to create a block widget
from a valid index.

GitHub Issue #7
Add C++/Qt Language Widgets