style: Explicitly defined methods.
style(*): With exemption to simple classes, replaced all using
declarations of methods into actual method definitions. These are all
constructors except for one method in the function C++ block which
overrides a parent method to make it private instead of public. This was
done so the design of these methods are in the design document and also
make it more clear in the source code itself.

GitHub Issue #7
Add C++/Qt Language Widgets