refactor: Updated unit model test class.
refactor(Test::UnitModel): Renamed the idle state to the none state
because none makes more sense as a state for a unit test result that has
never ran. Changed the background and foreground decoration roles for
the result column. The foreground is now black. The background is now
light colors of red, green, blue, and yellow for the respective states
of failed, passed, running, and none.

GitHub Issue #7
Add C++/Qt Language Widgets