refactor: Moved language enumerations.
refactor+(
  Cpp
  Cpp::Block
  CppQt
  CppQt::Block
): Moved enumerations which defined various aspects of language blocks
to the root namespace for each respective language. The only enumeration
that was not moved was the indexes for blocks. This was done because the
other enumerations defined fundamental aspects of each language that
should be global for that language.

refactor(
  Cpp::Block::Function
  CppQt::Block::Function
): Updated code to use new location of language enumerations.

test(tests): Updated all tests which used language enumerations to new
location of their respective languages.

GitHub Issue #10
Add C++/Qt Language Parsers