test: Added C++ router unit tests.
test(CppRouter): Added new unit test object which tests the C++ router.

fix(CppRouter): Fixed bug which caused the router to crash when adding
routes for the root namespace block because it had a Qt assert for the
name not being empty which it is for the root namespace. Fixed by
removing the check and the lambda where it was checked. The lambda was
removed because the first letter does not need to be capitalized for the
scope name.

fix(CppRouter): Fixed bug where a namespace would return an invalid and
unnecessary source code file route. This was caused because the
“has source” evaluator was incorrectly checking if a namespace block’s
class children have any source code. Fixed by removing the check on
class children. This caused the “has source” methods for namespace and
class blocks to be identical so they were merged, along with removing
the “has source” method for functions and simply transferring that code
in the new generic “has source” method.

GitHub Issue #10
Add C++/Qt Language Parsers