refactor: Overhauled block icon caching.
refactor!(
  Cpp::Class
  Cpp::Function
  Cpp::Property
  CppQt::Function
): Removed all getter methods for icons, replacing it with static
constant pointers to Qt icons used directly in the code where needed.
Added “initialize icons” static method to initialize all icons pointers
with Qt icon instances. It was does this way because the Qt icons must
be created after the Qt resources are loaded, which is not the cause if
the Qt icons are initialized globally.

refactor(
  Cpp::Language
  CppQt::Language
): Added call to new initialize icons methods for the class, function,
and property blocks in the constructor of their respective language
class. This insures all icons are initialized before any block object is
created.

GitHub Issue #10
Add C++/Qt Language Parsers