fix: Resolved copy paste bug.
fix(Command::Project::Insert): Discovered bug that caused a crash
whenever a block emitted its display icon or text changed signals. This
was caused due to passing the block pointer by reference with the class
instance pointer in the lambda functions that in turn called the project
slots connected to the signals. After the lambda functions are
connected, the block pointer is set to null, and therefore the lambda
functions pass a null pointer. This was resolved by passing the block
pointer by copy instead of reference to the lambda functions so the
block pointers are not overwritten with null inside the lamdba
functions.

fix(Cpp::Block::Base): Fixed minor bug where the root file name method
was not being called to return the root file name as it should have
been. This did not cause any issue because the string was correct but it
would have caused issues if the name was changed in the root file name
getter method.

test(TestModelProject): Added test that causes the copy paste bug. After
bug fix the new test now passes.

test(TestBlock): Added touch display text method that causes the block
to emit its display text changed signal. This was required for the new
test in the project model unit test class.

GitHub Issue #10
Add C++/Qt Language Parsers