fix: Resolved code controller bug.
fix(Controller::Code): While examining the code in the parse method, it
was noticed that there was a fatal bug in the lines dealing with
extraction of the source code version. It did not take into account
legacy source code… attempting to extract a version number regardless of
the first line matching the version syntax for not, and failing if it
could not get a number. This would cause any legacy file to fail. Fixed
by only extracting a version if the first line matches version syntax.
Also fixed legacy source code not starting on the first line.

GitHub Issue #10
Add C++/Qt Language Parsers