feature: Added EOL for parsing.
feature(AbstractParser): Added end of lines, or EOL, concept to the
parsing system. Now the active parser and all its parent parsers receive
a EOL indicator in their parse interface when the end of source code
lines has been reached. This will greatly benefit error checking while
parsing, allowing parsers to make sure they are expecting the source
code to end in their current state.

feature(Controller::Code): Implemented the end of lines concept to all
parsing code, passing the EOL value when the end of the code line list
is reached for the active parser and all parents. Changed internal
parsing method so it is no longer static and moved the code lines to the
class as a member variable along with the current file path being
parsed. This allows the internal parse method to throw a logical parse
error with the path to the source code file that caused it.

feature(
  Cpp::Parse::HeadParser
  Cpp::Parse::SourceParser
): Added code for handling the new EOL value.

test(
  TestCppParseHeadParser
  TestCppParseSourceParser
): Added tests to verify correct handling of EOL value. All new tests
pass.

GitHub Issue #10
Add C++/Qt Language Parsers