refactor: Merged base into abstract class.
refactor!(
  AbstractParser
  Controller::Code
  Cpp::Parse::BaseParser
): Removed the C++ base parser class, moving its protected insert code
method into the abstract parser class. The index key was changed from an
integer code key to a Qt string key, since the core library has no
concept of C++ code keys. Also added check for the given lines of code
which makes sure they are not empty. If the given lines are empty the
method does not insert anything.

refactor(
  Cpp::Parse::HeadParser
  Cpp::Parse::SourceParser
): Updated parent class back to the abstract parser. Updated use of
insert code method so the key string is passed and not the code key
integer.

feature(Cpp::Parse::SourceParser): Added parsing for version 1 which
ended up being trivial. This was trivial because there is no difference
between legacy and version 1, so the parse legacy method was used again
for version 1 parsing.

GitHub Issue #10
Add C++/Qt Language Parsers