refactor: Improved C++ parsers.
refactor(AbstractParser): Added optional block argument to the insert
code method, allowing the code to be inserted into a block other than a
parser object's block object.

refactor(AbstractParser): Simplified "on block destroyed" slot to throw
a logic error if the block destroyed matches the parser object's block
object.

refactor(Cpp::Parse::ClassParser): Added support for parsing multiple
class declarations by adding the ability to reset the parser to its
initial state. Modified constructor so it can take a namespace or class
block. If it is given a namespace block to will attempt to match any
child class contained in the namespace. If it is given a class it will
continue with its original behavior.

refactor(Cpp::Parse::FunctionParser): Renamed many private methods to be
shorter and more succinct. Moved some handling of function populating
into the constructor. The constructor now deals with determining if the
object's block object is a namespace or a class, passing the appropriate
arguments to the populate method. It also throws an exception if the
block object is not a class or namespace. Improved signature generation
by changing class name variables to scope with the 2 colon characters
always needed when inserting it before the function name.

fix(Cpp::Parse::FunctionParser): The constant flag was not correctly
being matched in function declarations, along with the no exception flag
not being accounted for at all. Both were resolved by moving where the
constant flag is expected in the regular expression and adding
additional code and regular expressions to account for no exception
flags.

docs(
  AbstractParser
  Cpp::Parse::ClassParser
  Cpp::Parse::FunctionParser
): Expanded upon documentation of classes and methods. Shortened
exception descriptions to simply state the types that can be thrown.

test(tests/CppParseClassParser): Updated first parsing test for legacy
and version 1 to handle multiple class declarations with the same parser
in the same source code file. A new version 1 parsing source code file
was required for its first pass due to version 1 parsing having more
branches that lead to ending the parsing of a class declaration. All
updated tests pass.

GitHub Issue #10
Add C++/Qt Language Parsers