fix: Resolved issues in C++ language.
fix(Cpp::Language): Resolved object casting issues in the create parser
interface. This was the same issue resolved in multiple C++ parsers,
where object casting was used to incorrectly determine a block's type.

refactor(
  Cpp::Language
  Cpp::Parse::FunctionParser
  Cpp::Parse::HeadParser
  Cpp::Parse::SourceParser
): Replaced logical parse exceptions with standard library logic errors
because this should be treated as internal errors that should never
happen as a result of parsing bad source code. The version is always
supplied by the C++ router, so if there is a bad version it is due to a
bug in the program that should be fixed.

GitHub Issue #10
Add C++/Qt Language Parsers