fix: Resolved C++ argument parsing.
fix(Cpp::Parse::HeadParser): Resolved issue in the de-tangle method
where parsing a list of function arguments would fail if any one of the
arguments had a type involving a template with more than one type in it
such as a Qt hash. The reason for this was because the string of
arguments were being split apart using the comma character, so the
commas inside a template would be seen as an argument separator. This
was resolved by adding more complexity to splitting up the arguments and
thereby allowing the code to ignore commas nested inside templates. The
"to type" method was also upgraded additional argument validation and
moving one of its checks to the de-tangle method.

GitHub Issue #10
Add C++/Qt Language Parsers