fix: Resolved multiple issues.
fix(Controller::Code): Resolved two issues in the parse method. When
catching a logical parse error and adding the line number information to
it, the string was not correctly formatted resulting in mangled output.
Because the parse method is recursive, multiple line numbers would be
appended to the error message as the stack calls of parse are destroyed.
Added static caught line variable to report the correct line where the
error is thrown.

fix(Cpp::Router): Resolved issue in the add routes method. When
recursively calling said method the new scope was not being passed to
the next call, instead the old scope was being passed resulting in
previous scope names never being added.

test(CppRouter): Added test for bug found in the C++ router. After
fixing the bug the new test passes.

GitHub Issue #10
Add C++/Qt Language Parsers